+++ src/components/Instructions/MathematicalToolsRow.tsx
@@ -1,9 +1,9 @@
 import styled from 'styled-components';
-import AnimatedLottieIcon from './AnimatedLottieIcon';
-import rotateToolAnimation from '../assets/icons-animation/rotate-tool.json';
-import horizontalStretchAnimation from '../assets/icons-animation/horizontal-stretch-tool.json';
-import cutAnimation from '../assets/icons-animation/cut-tool.json';
-import combineAnimation from '../assets/icons-animation/combine-tool.json';
+import AnimatedLottieIcon from '../AnimatedLottieIcon';
+import rotateToolAnimation from '../../assets/icons-animation/rotate-tool.json';
+import horizontalStretchAnimation from '../../assets/icons-animation/horizontal-stretch-tool.json';
+import cutAnimation from '../../assets/icons-animation/cut-tool.json';
+import combineAnimation from '../../assets/icons-animation/combine-tool.json';
 
 function MathematicalToolsRow() {
   return (
