+++ src/levels.json
@@ -16,14 +16,86 @@
       [{ "x": 0, "y": 2, "width": 5, "height": 4, "valid": true }]
     ],
     "pieces": [
-      { "width": 2, "height": 1 },
-      { "width": 3, "height": 3 },
       { "width": 2, "height": 3 },
-      { "width": 3, "height": 3 }
+      { "width": 3, "height": 3 },
+      { "width": 3, "height": 3 },
+      { "width": 2, "height": 1 }
     ]
   },
   {
     "id": 2,
+    "dimensions": { "width": 7, "height": 6 },
+    "boardSections": [
+      [
+        {
+          "x": 0,
+          "y": 0,
+          "width": 5,
+          "height": 3,
+          "valid": true
+        },
+        { "x": 5, "y": 0, "width": 2, "height": 3, "valid": false }
+      ],
+      [{ "x": 0, "y": 3, "width": 7, "height": 3, "valid": true }]
+    ],
+    "pieces": [
+      { "width": 2, "height": 1 },
+      { "width": 3, "height": 3 },
+      { "width": 2, "height": 3 },
+      { "width": 3, "height": 3 },
+      { "width": 2, "height": 5 }
+    ]
+  },
+  {
+    "id": 3,
+    "dimensions": { "width": 7, "height": 7 },
+    "boardSections": [
+      [
+        {
+          "x": 0,
+          "y": 0,
+          "width": 7,
+          "height": 7,
+          "valid": true
+        }
+      ]
+    ],
+    "pieces": [
+      { "width": 4, "height": 5 },
+      { "width": 2, "height": 2 },
+      { "width": 3, "height": 4 },
+      { "width": 2, "height": 3 },
+      { "width": 7, "height": 1 }
+    ]
+  },
+  {
+    "id": 4,
+    "dimensions": { "width": 7, "height": 8 },
+    "boardSections": [
+      [
+        {
+          "x": 0,
+          "y": 0,
+          "width": 7,
+          "height": 2,
+          "valid": true
+        }
+      ],
+      [
+        { "x": 0, "y": 2, "width": 4, "height": 6, "valid": true },
+        { "x": 4, "y": 2, "width": 3, "height": 6, "valid": false }
+      ]
+    ],
+    "pieces": [
+      { "width": 2, "height": 5 },
+      { "width": 2, "height": 2 },
+      { "width": 3, "height": 4 },
+      { "width": 2, "height": 3 },
+      { "width": 3, "height": 2 }
+    ]
+  },
+  {
+    "id": 5,
     "dimensions": { "width": 7, "height": 8 },
     "boardSections": [
       [{ "x": 0, "y": 0, "width": 7, "height": 8, "valid": true }]
@@ -40,7 +112,7 @@
     ]
   },
   {
-    "id": 3,
+    "id": 6,
     "dimensions": { "width": 15, "height": 14 },
     "boardSections": [
       [
@@ -67,7 +139,7 @@
     ]
   },
   {
-    "id": 4,
+    "id": 7,
     "dimensions": { "width": 10, "height": 11 },
     "boardSections": [
       [
@@ -92,7 +164,7 @@
     ]
   },
   {
-    "id": 5,
+    "id": 8,
     "dimensions": { "width": 18, "height": 12 },
     "boardSections": [
       [
@@ -119,7 +191,7 @@
     ]
   },
   {
-    "id": 6,
+    "id": 9,
     "dimensions": { "width": 11, "height": 8 },
     "boardSections": [
       [
@@ -149,7 +221,7 @@
     ]
   },
   {
-    "id": 7,
+    "id": 10,
     "dimensions": { "width": 10, "height": 10 },
     "boardSections": [
       [
@@ -195,7 +267,7 @@
     ]
   },
   {
-    "id": 8,
+    "id": 11,
     "dimensions": { "width": 10, "height": 14 },
     "boardSections": [
       [
@@ -248,7 +320,7 @@
     ]
   },
   {
-    "id": 9,
+    "id": 12,
     "dimensions": { "width": 17, "height": 19 },
     "boardSections": [
       [
@@ -311,7 +383,7 @@
     ]
   },
   {
-    "id": 10,
+    "id": 13,
     "dimensions": { "width": 20, "height": 20 },
     "boardSections": [
       [
@@ -340,6 +412,162 @@
       { "width": 8, "height": 1 },
       { "width": 4, "height": 7 },
       { "width": 5, "height": 3 }
+    ]
+  },
+  {
+    "id": 14,
+    "dimensions": { "width": 16, "height": 18 },
+    "boardSections": [
+      [
+        {
+          "x": 0,
+          "y": 0,
+          "width": 16,
+          "height": 5,
+          "valid": true
+        }
+      ],
+      [
+        {
+          "x": 0,
+          "y": 5,
+          "width": 4,
+          "height": 3,
+          "valid": false
+        },
+        {
+          "x": 4,
+          "y": 5,
+          "width": 8,
+          "height": 3,
+          "valid": true
+        },
+        {
+          "x": 12,
+          "y": 5,
+          "width": 4,
+          "height": 3,
+          "valid": false
+        }
+      ],
+      [
+        {
+          "x": 0,
+          "y": 8,
+          "width": 16,
+          "height": 10,
+          "valid": true
+        }
+      ]
+    ],
+    "pieces": [
+      { "width": 7, "height": 2 },
+      { "width": 1, "height": 6 },
+      { "width": 8, "height": 5 },
+      { "width": 3, "height": 1 },
+      { "width": 3, "height": 4 },
+      { "width": 6, "height": 2 },
+      { "width": 4, "height": 3 },
+      { "width": 4, "height": 6 },
+      { "width": 5, "height": 5 },
+      { "width": 9, "height": 4 },
+      { "width": 3, "height": 8 },
+      { "width": 8, "height": 4 },
+      { "width": 2, "height": 2 },
+      { "width": 4, "height": 5 }
+    ]
+  },
+  {
+    "id": 15,
+    "dimensions": { "width": 20, "height": 20 },
+    "boardSections": [
+      [
+        {
+          "x": 0,
+          "y": 0,
+          "width": 20,
+          "height": 20,
+          "valid": true
+        }
+      ]
+    ],
+    "pieces": [
+      { "width": 7, "height": 8 },
+      { "width": 5, "height": 6 },
+      { "width": 8, "height": 4 },
+      { "width": 3, "height": 4 },
+      { "width": 6, "height": 2 },
+      { "width": 8, "height": 6 },
+      { "width": 5, "height": 5 },
+      { "width": 9, "height": 4 },
+      { "width": 3, "height": 12 },
+      { "width": 1, "height": 10 },
+      { "width": 8, "height": 6 },
+      { "width": 2, "height": 2 },
+      { "width": 8, "height": 1 },
+      { "width": 4, "height": 7 },
+      { "width": 5, "height": 3 }
+    ]
+  },
+  {
+    "id": 16,
+    "dimensions": { "width": 14, "height": 14 },
+    "boardSections": [
+      [
+        {
+          "x": 0,
+          "y": 0,
+          "width": 14,
+          "height": 5,
+          "valid": true
+        }
+      ],
+      [
+        {
+          "x": 0,
+          "y": 5,
+          "width": 5,
+          "height": 4,
+          "valid": true
+        },
+        {
+          "x": 5,
+          "y": 5,
+          "width": 4,
+          "height": 4,
+          "valid": false
+        },
+        {
+          "x": 9,
+          "y": 5,
+          "width": 5,
+          "height": 4,
+          "valid": true
+        }
+      ],
+      [
+        {
+          "x": 0,
+          "y": 9,
+          "width": 14,
+          "height": 5,
+          "valid": true
+        }
+      ]
+    ],
+    "pieces": [
+      { "width": 7, "height": 4 },
+      { "width": 5, "height": 5 },
+      { "width": 8, "height": 3 },
+      { "width": 3, "height": 3 },
+      { "width": 6, "height": 2 },
+      { "width": 2, "height": 4 },
+      { "width": 6, "height": 3 },
+      { "width": 2, "height": 4 },
+      { "width": 6, "height": 2 },
+      { "width": 4, "height": 3 },
+      { "width": 5, "height": 4 },
+      { "width": 1, "height": 4 }
     ]
   }
 ]
